new movement

+++ TowerDefense/TowerDefense/Game.h
@@ -45,7 +45,7 @@ public:
 		int i, float y, float x, sf::Sprite &lifeEnemySprite, int &playerLife, sf::Text &lebenText,
 		vector<GameArea*>& playingField, vector<int> &path);
 	void UpdateEnemyLifeBar(std::vector<BasicEnemy *> * enemyActiveVector, int i, int &punkteZahl,
-		int &gold, float &x, float &y, sf::Sprite &lifeEnemySprite, sf::Texture &hundredLifeTexture,
+		int &gold, int &x, int &y, sf::Sprite &lifeEnemySprite, sf::Texture &hundredLifeTexture,
 		sf::Texture &eightyLifeTexture, sf::Texture &sixtyLifeTexture, sf::Texture &fortyLifeTexture,
 		sf::Texture &twentyLifeTexture, sf::Texture &tenLifeTexture, sf::Text &punktZahlText);
 	void ShowGameOverScreen(sf::Font &font, sf::Color &color, sf::Texture &backgroundTexture, sf::Sprite &backgroundSprite, sf::String &playerInput, sf::Text &playerText);
